fix(web): preserve embedded field focus and simplify page components
