+++ frontend/apps/web/src/components/agents/AgentProfileNameHeading.tsx
@@ -59,8 +59,9 @@ export function AgentProfileNameHeading({
               autoFocus
               aria-label="Profile name"
               required
+              variant="embedded"
               value={nameDraft}
-              className="border-border focus-visible:border-border type-title -mx-1 h-auto w-full min-w-0 max-w-96 rounded-none border-0 border-b px-1 py-0 text-2xl shadow-none focus-visible:ring-0 sm:h-auto md:text-2xl dark:bg-transparent"
+              className="border-border focus-visible:border-ring type-title -mx-1 h-auto w-full min-w-0 max-w-96 rounded-none border-b px-1 py-0 text-2xl sm:h-auto md:text-2xl"
               onChange={(event) => {
                 setNameDraft(event.target.value)
               }}
